+++ surpyval/univariate/parametric/distributions/gumbel.py
@@ -223,18 +223,18 @@ class Gumbel_(OptimisedFitMixin, ParametricFitter):
         """
         return np.exp((x - mu) / sigma)
 
-    def qf(self, p, mu, sigma):
+    def qf(self, u, mu, sigma):
         r"""
 
         Quantile function for the Gumbel Distribution:
 
         .. math::
-            q(p) = \mu + \sigma\ln\left ( -\ln\left ( 1 - p \right ) \right )
+            q(u) = \mu + \sigma\ln\left ( -\ln\left ( 1 - u \right ) \right )
 
         Parameters
         ----------
 
-        p : numpy array or scalar
+        u : numpy array or scalar
             The percentiles at which the quantile will be calculated
         mu : numpy array like or scalar
             The location parameter(s) of the distribution
@@ -245,17 +245,17 @@ class Gumbel_(OptimisedFitMixin, ParametricFitter):
         -------
 
         q : scalar or numpy array
-            The quantiles for the Gumbel distribution at each value p.
+            The quantiles for the Gumbel distribution at each value u.
 
         Examples
         --------
         >>> import numpy as np
         >>> from surpyval import Gumbel
-        >>> p = np.array([0.1, 0.3, 0.5])
-        >>> Gumbel.qf(p, 3, 2)
+        >>> u = np.array([0.1, 0.3, 0.5])
+        >>> Gumbel.qf(u, 3, 2)
         array([-1.50073465,  0.93813913,  2.26697416])
         """
-        return mu + sigma * (np.log(-np.log1p(-p)))
+        return mu + sigma * (np.log(-np.log1p(-u)))
 
     def mean(self, mu, sigma):
         r"""
@@ -301,8 +301,8 @@ class Gumbel_(OptimisedFitMixin, ParametricFitter):
     def log_ff(self, x, mu, sigma):
         return np.log(-np.expm1(-self.Hf(x, mu, sigma)))
 
-    def moment(self, n, mu, sigma):
-        return gumbel_l.moment(n, loc=mu, scale=sigma)
+    def moment(self, m, mu, sigma):
+        return gumbel_l.moment(m, loc=mu, scale=sigma)
 
     def entropy(self, mu, sigma):
         r"""
@@ -336,8 +336,8 @@ class Gumbel_(OptimisedFitMixin, ParametricFitter):
         """
         return np.log(sigma) + euler_gamma + 1
 
-    def mpp_x_transform(self, x, gamma=0):
-        return x - gamma
+    def mpp_x_transform(self, x):
+        return x
 
     def mpp_y_transform(self, y, *params):
         mask = (y == 0) | (y == 1)